backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
asyncpg==0.27.0
alembic==1.12.0
pydantic==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
aiosmtplib==1.1.6
httpx==0.24.1
cryptography==41.0.3
reportlab==4.1.0
qrcode==7.4.2
pillow==10.0.0
psycopg2-binary==2.9.7

pydantic-settings==2.0.3
